feat: implement MongoDB integration and refactor services for player and match management
